refactor(wordle history): componentize
