+++ src/app/(pages)/games/wordleverse/history/_actions/getHistory.js
@@ -1,53 +1,19 @@
 "use server";
 
 import { getCurrentUser } from "@/lib/auth";
-import prisma from "@/lib/prisma";
+import { getHistoryFromDB } from "../_lib/storage/database";
 
-import { getAvailableDates } from "./getAvailableDates";
-
+/**
+ * Server action to get user's game history
+ * This is a wrapper around the database storage function
+ * @param {Object} options - Options for fetching history
+ * @returns {Object} The history data or error
+ */
 export const getHistory = async (options = {}) => {
-   const user = await getCurrentUser();
-    if (!user) {
-      return { error: "Unauthorized", status: 401 };
-    }
-    const userId = user.id;
+  const user = await getCurrentUser();
+  if (!user) {
+    return { error: "Unauthorized", status: 401 };
+  }
   
-    try {
-      // Get user's game history with streak information
-      const games = await prisma.wordleHistory.findMany({
-        where: {
-          userId,
-        },
-        orderBy: {
-          date: "desc",
-        },
-      });
-    
-      // Get user's streak information
-      const user = await prisma.user.findUnique({
-        where: { id: userId },
-        select: { streak: true, maxStreak: true },
-      });
-
-      const history = {
-        games,
-        streak: user?.streak || 0,
-        maxStreak: user?.maxStreak || 0,
-      }
-
-      // If requested, include available dates
-      const includeAvailableDates = options.includeAvailableDates === true;
-      if (includeAvailableDates) {
-        const availableDates = await getAvailableDates(userId);
-        return {
-          ...history,
-          availableDates
-        };
-      }
-      
-      return history;
-    } catch (error) {
-      console.error("Error getting history:", error);
-      return { error: "Failed to get history", status: 500 };
-    }
-}
+  return await getHistoryFromDB(options);
+};
